LPD-18488 Isolate useUndoRedoActions to test it better
