+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/undo/Undo.js
@@ -4,28 +4,11 @@
  */
 
 import ClayButton, {ClayButtonWithIcon} from '@clayui/button';
-import PropTypes from 'prop-types';
 import React from 'react';
 
-import {useDispatch, useSelector} from '../../contexts/StoreContext';
-import redo from '../../thunks/redo';
-import undo from '../../thunks/undo';
+import {useSelector} from '../../contexts/StoreContext';
 import UndoHistory from './UndoHistory';
-
-export function useUndoRedo() {
-	const dispatch = useDispatch();
-	const store = useSelector((state) => state);
-
-	const onUndo = () => {
-		dispatch(undo({store}));
-	};
-
-	const onRedo = () => {
-		dispatch(redo({store}));
-	};
-
-	return {onRedo, onUndo};
-}
+import useUndoRedoActions from './useUndoRedoActions';
 
 export function useDisabledUndo() {
 	const undoHistory = useSelector((state) => state.undoHistory);
@@ -42,7 +25,7 @@ export function useDisabledRedo() {
 export default function Undo() {
 	const disabledRedo = useDisabledRedo();
 	const disabledUndo = useDisabledUndo();
-	const {onRedo, onUndo} = useUndoRedo();
+	const {onRedo, onUndo} = useUndoRedoActions();
 
 	return (
 		<>
@@ -76,8 +59,3 @@ export default function Undo() {
 		</>
 	);
 }
-
-Undo.propTypes = {
-	onRedo: PropTypes.func,
-	onUndo: PropTypes.func,
-};
